qcacld-3.0: Strip additional 3 bytes allocated for auth wrapped data
When FILS erp keys are not available, EAP initiate packet will not
be sent to the AP so FILS authentication wrapped data IE is not
required, but the driver is adding extra 3 bytes for wrapped data
even if wrapped data is not included. This will cause issues with
the FILS AP.

Don't allocate bytes for wrapped data if ERP keys are not available.

Change-Id: I9bd8033048829eeaf63ce123129e64f2685df04f
CRs-Fixed: 2732852
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>