qcacld-3.0: Add high priority work to update netdev features
In certain special scenarios, we need to update the
netdev features of a particular adapter eg: enabling or
disabling tso or hardware checksum incase of legacy clients.

Such an update of the netdev features is to be done only
after the connection, since the connection mode plays an
important role in identifying the features which are to be
updated.

The netdev features can be updated only while holding the
rtnl lock. Also taking into account that the protocol layer,
where we get the idea of the mode of connection, is under its
own locks, updating the netdev features while holding rtnl lock
can often lead to deadlock.

Hence to mitigate the above mentioned deadlock issue, we always
queue a high-priority work to update the netdev features for
every adapter. The need for high-priority work arises from the need
to maintain the roaming KPIs.

Change-Id: Ie9218b08931dea099c6194efa7434d888b849e8d
CRs-Fixed: 2735636
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>