qcacld-3.0: Print current RSSI of AP after roam scan is complete
In wma_get_trigger_detail_str() driver prints the roam trigger
detail for Low rssi and Periodic scan triggers, the RSSI value of
the current AP is the value before roam scan is triggered. But
the firmware considers the rssi of the current AP after roam scan
is complete to calculate the next RSSI threshold. So there could
be mismatch in the current rssi value and next rssi threshold
when roam_info->rssi_trig_data.threshold is used to print the
current AP rssi.

Use roam_info->current_rssi to print the current connected AP
rssi in the roam trigger logging.

Change-Id: Ib154bb3cbdfe13dbcdd8c8830cdafb6daefa34b5
CRs-Fixed: 2739346
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>