qcacld-3.0: Introduce CONFIG_WLAN_TRACE_HIDE_MAC_ADDRESS
MAC address is of the format (XX:XX:XX:XX:XX:XX),
This feature is to hide logging 2nd, 3rd and 4th octet of the
MAC address in the logs.

To enable this feature set CONFIG_WLAN_TRACE_HIDE_MAC_ADDRESS to y.
For example, if this feature is enabled, then
mac address ae:45:9c:f1:c0:98 shall be printed as ae:**:**:**:c0:98

If this feature is disabled, then QDF_MAC_ADDR_FMT translates to
"%pM" and by default this feature disabled.

Change-Id: Ib4086c1390c3ecafa637616954939f7d2c9c0e5b
CRs-Fixed: 2752355
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>