qcacld-3.0: fix an out-of-bounds read issue
Size of rsn_caps/mlme_ht_capabilities_info/mlme_ht_ext_cap_info are 16bit,
mlme_ht_param_info is 8bit, it casts the pointer type to (uint32_t *) to
get a 32bit value, which results in out-of-bounds read.
The fix is to cast according to the size of the original type.

CRs-Fixed: 2752734
Change-Id: I20131b8539157f30471723d7211fbf8b209c494f
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>