qca-wifi: Allow RootAP side devices to be re-learnt on Primary station
The change allows RootAP side devices to be re-learn on Primary
station. This is to take care of a schenario when secondary
station vap comes up first and RootAP side devices are learnt
on secondary station and the when primary station comes up
all the devices have to be re-learnt on Primary station.

Change-Id: If2317eab94edb145fbbff27046c7a5bc456a3302
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>