qcacld-3.0: Wait for hw mode change done for vdev delete
Vdev delete will purge the serialization queue for the vdev.
If the hw mode change request is flushed without waiting for
the fw completion event, it will cause issues when the fw event
coming up later and no active serialization command in queue.
In one case, the “hw_mode_change_in_progress” is not cleared
in this condition. And finally causes the new Connection req
failed to wait for hw mode change event because there is no
hw change request actually.

Fix by check and wait for hw mode response if any hw mode change
is in progress.

Change-Id: I007643f48ff1d35fbdb2c9ca9a3127bbd66d7dec
CRs-Fixed: 2760395
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>