qcacld-3.0: Save low latency connection count in DP vdev
The 11g and 11a connections are latency critical and
hence tx coalescing cannot be applied on traffic
running for these connections.

Detect the latency critical connections and save the
number of latency critical connections in DP vdev.
This information can be used by the Software latency
manager to skip tx register write coalescing when there
is a packet transmission for low latency connections.

Change-Id: I2ebf40e670cc3139b2bc41a690c79ebb31c7b9c7
CRs-Fixed: 2769026
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>