qcacmn: Search opclass from global opclass also
While querying for the channel width, the opclass is
extracted from the country opclass. If opclass is not
found then 20Mhz channel width is considered for connection.
This fails the cert test case when AP sends the global
opclass in ECSA

Change is done to search in the global opclass if the
opclass is not found in the country specific opclass.

CRs-Fixed: 2763702
Change-Id: If8c962e1e164cfccda13f19cbf925499c9cc0394
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>