qcacmn: Avoid peer free notification for self peer
The peer free notification is sent for all peers
including self peer. The primiary intent of using this notification
is to make sure all the connected peers are disconnected.

Also when the peer free notification is sent, the peer is already
freed and we cannot identify if the peer free is for self peer

The self peer free notification for STA is causing issues
where BSS peer delete is skipped as part of bringing down the VDEV SM.

CRs-Fixed: 2755824
Change-Id: I276177788a51dcdd948caeb734eaa896edf520dd
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>